sampleproject/requirements.txt
fastapi
uvicorn
motor
redis
python-dotenv
databases[postgresql]
psycopg2-binary
pika